+++ changelog.md
@@ -7,3 +7,4 @@
   convenience functions to change file permissions using Unix like octal file permissions.
 - Added module `scripting` providing `withDir` to switch the directory temporarily. This
   was previously only available in the `nimscript` module.
+- Added `openArray[char]` overloads for `filepermissions.chmod` and `filepermissions.toFilePermissions`.